add product and fix user login register

+++ src/app/products/products.controller.ts
@@ -1,20 +1,34 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { Role } from '@/security/role.decorator';
+import { Roles, User } from '@/security/user.decorator';
 
 @Controller('products')
 export class ProductsController {
   constructor(private readonly productsService: ProductsService) {}
 
   @Post()
-  create(@Body() createProductDto: CreateProductDto) {
-    return this.productsService.create(createProductDto);
+  @Role(Roles.org)
+  create(
+    @Body() createProductDto: CreateProductDto,
+    @User('orgId') orgId: number,
+  ) {
+    return this.productsService.create(orgId, createProductDto);
   }
 
   @Get()
-  findAll() {
-    return this.productsService.findAll();
+  findAll(@User('orgId') orgId: number) {
+    return this.productsService.findAll(orgId);
   }
 
   @Get(':id')
@@ -23,12 +37,20 @@ export class ProductsController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
+  @Role(Roles.org)
+  async update(
+    @Param('id') id: string,
+    @Body() updateProductDto: UpdateProductDto,
+    @User('orgId') orgId: number,
+  ) {
+    await this.productsService.checkOwnership(+id, orgId);
     return this.productsService.update(+id, updateProductDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @Role(Roles.org)
+  async remove(@Param('id') id: string, @User('orgId') orgId: number) {
+    await this.productsService.checkOwnership(+id, orgId);
     return this.productsService.remove(+id);
   }
 }
